+++ CHANGELOG.md
@@ -38,6 +38,7 @@ Unless otherwise specified, any version comparison below is the comparison of se
 - Fixed a bug where the player dying in a non-overworld dimension could not be properly teleported back to the spawn point in overworld.
 - Fixed a NPE when interacting bed block in a non-overworld dimension.
 - Fixed a bug where chunks sent to the client sometimes did not display immediately.
+- Fixed a NPE which rarely happens in the light engine.
 
 ### Removed